# 3-technical/3.1-system-foundation/architecture/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AKAIUNSAN Attendance + Payroll API
  version: 1.0.0
  description: |
    REST API for AKAIUNSAN attendance and payroll systems (PRD-EPIC-002, Phase 0).

    **Auth:** Bearer JWT (15-min TTL) + refresh token cookie (30 days).
    **Internal:** `X-Internal-API-Key` header (service-to-service only).
    **Rate limits:** Production API instances enforce a global 100 requests/minute limit; auth endpoints add OTP/challenge abuse controls.
    **Time:** ISO 8601 UTC; VN time (UTC+7) at presentation layer.
    **Money:** VNĐ as string (Decimal); never float.
    **Errors:** `{ error: { code, message, details? } }`

    Full schemas live in this file. Full business rules in
    [Domain Specs](../domain-specs.md). Architecture in
    [System Design](../../design-standards/system-design.md).

    Code conventions in [Coding Standards](../../design-standards/coding-standards.md).

  contact:
    name: AKAIUNSAN Engineering
    email: engineering@ak.local
...
security:
  - bearerAuth: []
paths:
...
  /v1/employees:
    get:
      tags: [identity]
      summary: List employees
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/LimitParam' }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedEmployees' }
    post:
      tags: [identity]
      summary: Create employee (BO/system admin; permission-gated)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeCreate' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/projects:
    get:
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, paused, ended] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
...
  /v1/shifts:
    get:
      tags: [shifts]
      summary: List shift templates
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Shift' }
    post:
      tags: [shifts]
      summary: Create shift template
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftCreate' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shift' }
  /v1/shifts/assignments:
    get:
      tags: [shifts]
      summary: List shift assignments
      parameters:
        - in: query
          name: employeeId
          schema: { type: string, format: uuid }
        - in: query
          name: projectId
          schema: { type: string, format: uuid }
        - in: query
          name: from
          description: Inclusive shift-assignment work date.
          schema: { type: string, format: date }
        - in: query
          name: to
          description: Inclusive shift-assignment work date.
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ShiftAssignment' }
    post:
      tags: [shifts]
      summary: Assign shift to employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftAssignmentCreate' }
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShiftAssignment' }
        '422': { $ref: '#/components/responses/BusinessRuleViolation' }

  # ===== ATTENDANCE =====
...
  /internal/attendance:
    get:
      tags: [internal]
      summary: Attendance data for payroll (internal only)
      security:
        - internalApiKey: []
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: employeeId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: from
          required: true
          description: Inclusive assignment work date; the inclusive range may contain at most 366 calendar days.
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          description: Inclusive assignment work date; must be on or after `from`.
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/InternalAttendanceRecord' }
        '400':
          description: Invalid calendar date, reversed range, or range longer than 366 days
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (15-min TTL)
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Shared secret, rotated quarterly
  parameters:
    PageParam:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Resource already exists or state violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BusinessRuleViolation:
      description: Domain rule violated (e.g., GPS outside geofence)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
    Employee:
      type: object
      required: [id, userId, employeeCode, fullName, hireDate, baseSalary, salaryType, status]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        employeeCode: { type: string, example: NV001 }
        fullName: { type: string, example: 'Nguyễn Văn A' }
        dateOfBirth: { type: [string, 'null'], format: date }
        hireDate: { type: string, format: date }
        baseSalary: { type: string, description: 'VNĐ as string Decimal, e.g. 10000000.00' }
        salaryType: { type: string, enum: [monthly, hourly] }
        hourlyRate: { type: [string, 'null'] }
        bankAccount: { type: [string, 'null'] }
        bankName: { type: [string, 'null'] }
        idNumber: { type: [string, 'null'] }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        user:
          type: object
          required: [id, phone, role, status]
          properties:
            id: { type: string, format: uuid }
            email: { type: [string, 'null'], format: email }
            phone: { type: string }
            role: { type: string, enum: [employee] }
            status: { type: string, enum: [active, inactive, suspended] }
        temporaryPassword: { type: string, description: Present only on employee creation }
    EmployeeSupervisorView:
      type: object
      required: [id, userId, employeeCode, fullName, status, user]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        employeeCode: { type: string }
        fullName: { type: string }
        status: { type: string, enum: [active, inactive] }
        user:
          type: object
          required: [id, phone, role, status]
          properties:
            id: { type: string, format: uuid }
            email: { type: [string, 'null'], format: email }
            phone: { type: string }
            role: { type: string, enum: [employee] }
            status: { type: string, enum: [active, inactive, suspended] }
        temporaryPassword: { type: string, description: Present only on employee creation }
...
    EmployeeResponse:
      oneOf:
        - $ref: '#/components/schemas/Employee'
        - $ref: '#/components/schemas/EmployeeSupervisorView'
    EmployeeCreate:
      type: object
      required: [phone, fullName, hireDate, baseSalary, salaryType]
      properties:
        phone: { type: string, pattern: '^\+84[0-9]{9}$' }
        fullName: { type: string }
        employeeCode: { type: string }
        dateOfBirth: { type: string, format: date }
        hireDate: { type: string, format: date, description: Must not be in the future in the Vietnam business calendar. }
        baseSalary: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        salaryType: { type: string, enum: [monthly, hourly] }
        hourlyRate: { type: string, pattern: '^\d+(?:\.\d{1,2})?$' }
        bankAccount: { type: string }
        bankName: { type: string }
        idNumber: { type: string }
        temporaryPassword: { type: string, minLength: 8 }
      allOf:
        - if:
            properties:
              salaryType: { const: hourly }
            required: [salaryType]
          then:
            required: [hourlyRate]
            properties:
              hourlyRate:
                type: string
                pattern: '^(?:[1-9]\d*(?:\.\d{1,2})?|0\.(?:0[1-9]|[1-9]\d?))$'
...
    PaginatedEmployees:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/EmployeeResponse' }
        pagination: { $ref: '#/components/schemas/Pagination' }

    # ===== Project =====
    Project:
      type: object
      required: [id, code, name, clientName, address, latitude, longitude, geofenceRadiusMeters, status]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PRJ001 }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, default: 100 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: [string, 'null'], format: date }
        status: { type: string, enum: [active, paused, ended] }
    ProjectCreate:
      type: object
      required: [code, name, clientName, address, latitude, longitude, contractStartDate]
      properties:
        code: { type: string }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, default: 100 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: string, format: date }
...
    Shift:
      type: object
      required: [id, name, startTime, endTime, breakMinutes, lateThresholdMinutes, isOvernight]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 'Ca sáng' }
        startTime: { type: string, example: '06:00' }
        endTime: { type: string, example: '14:00' }
        breakMinutes: { type: integer, default: 60 }
        lateThresholdMinutes: { type: integer, default: 15 }
        isOvernight: { type: boolean, default: false }
        color: { type: string, example: '#FF5733' }
    ShiftCreate:
      type: object
      required: [name, startTime, endTime]
      properties:
        name: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        breakMinutes: { type: integer }
        lateThresholdMinutes: { type: integer }
        isOvernight: { type: boolean }
        color: { type: string }
    ShiftAssignment:
      type: object
      required: [id, employeeId, projectId, shiftId, date, status]
      properties:
        id: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        shiftId: { type: string, format: uuid }
        date: { type: string, format: date }
        status: { type: string, enum: [scheduled, checked_in, checked_out, completed, missed, cancelled] }
    ShiftAssignmentCreate:
      type: object
      required: [employeeId, projectId, shiftId, date]
      properties:
        employeeId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        shiftId: { type: string, format: uuid }
        date: { type: string, format: date }
        notes: { type: string }
...
    InternalAttendanceRecord:
      type: object
      additionalProperties: false
      required: [id, shiftAssignmentId, employeeId, workDate, status]
      properties:
        id: { type: string, format: uuid }
        shiftAssignmentId: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        workDate: { type: string, format: date, description: Shift-assignment business date used for payroll period and holiday/weekend classification. }
        checkInAt: { type: [string, 'null'], format: date-time }
        checkOutAt: { type: [string, 'null'], format: date-time }
        status: { type: string, enum: [present, late, early_leave, half_day, absent, on_leave, holiday] }
        totalMinutesWorked: { type: [integer, 'null'] }
        overtimeMinutes: { type: [integer, 'null'] }
        lateMinutes: { type: [integer, 'null'] }
...